requirements.txt
rfdetr[train]>=1.6.0
supervision>=0.21.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
faiss-cpu>=1.8.0
easyocr>=1.7.1
rapidfuzz>=3.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
pandas>=2.2.0
Pillow>=10.3.0
numpy>=1.26.0
requests>=2.31.0
python-multipart>=0.0.9
